Moved big functions to independent files
