+++ src/BlockAnalysis.ts
@@ -1,21 +1,19 @@
 export interface BlockAnalysis {
-
     localVariables: {
         name: string;
         // location: CodeLocation
-        
+
         /**
          * means that it shares the name of a variable in and earlier block
          */
-        shadows: boolean
-    }[]
+        shadows: boolean;
+    }[];
 
     referencedVariables: {
         name: string;
         // location: CodeLocation
         block: number; // blocks above the current block parent number
-    }
-
+    };
 }
 
 /*
