+++ src/analyzeCode.ts
@@ -1,9 +1,11 @@
-import ts, { BinaryOperatorToken } from "typescript";
+import ts from "typescript";
 import { createProgramFromFiles } from "./createProgramFromFiles";
 import { VariableStatementAnalysis } from "./VariableStatementAnalysis";
 import { getNodePosition } from "./getNodePosition";
 import { Stack } from "./stack";
-import {CodeAnalysis} from "./CodeAnalysis"
+import { CodeAnalysis } from "./CodeAnalysis";
+import { processExpression } from "./coreAnalyzer";
+import { editVariables } from "./editVariable";
 
 /**
  * visit top level nodes and retrieve all VariableStatements.
@@ -39,7 +41,7 @@ export function analyzeCode(code: string): CodeAnalysis {
         if (ts.isVariableStatement(node)) {
             const variableType = node.declarationList.getChildAt(0);
             //calculate the value of that variable and add it to the variables array
-            let declarationsList = node.declarationList.declarations;
+            const declarationsList = node.declarationList.declarations;
             declarationsList.forEach(function (expression) {
                 const variableValue = processExpression(
                     //get the expression of the variable declaration
@@ -103,310 +105,11 @@ export function analyzeCode(code: string): CodeAnalysis {
     // iterate through source file searching for variable statements
     visitNodeRecursive(sourceFile, visitVariableStatement);
 
-
     // TODO: actually do block analysis
     return {
         variableStatementAnalysis: detectedVariableStatements,
         blockAnalysis: [],
     };
-}
-
-const operations = new Map<ts.SyntaxKind, (a: number, b: number) => number>([
-    [ts.SyntaxKind.PlusToken, (a: number, b: number) => a + b],
-    [ts.SyntaxKind.MinusToken, (a: number, b: number) => a - b],
-    [ts.SyntaxKind.AsteriskToken, (a: number, b: number) => a * b],
-    [ts.SyntaxKind.AsteriskAsteriskToken, (a: number, b: number) => a ** b],
-    [ts.SyntaxKind.SlashToken, (a: number, b: number) => a / b],
-    [ts.SyntaxKind.PercentToken, (a: number, b: number) => a % b],
-]);
-
-const numberBooleanOperations = new Map<
-    ts.SyntaxKind,
-    (a: number, b: number) => boolean
->([
-    [ts.SyntaxKind.EqualsEqualsEqualsToken, (a: number, b: number) => a === b],
-    [
-        ts.SyntaxKind.ExclamationEqualsEqualsToken,
-        (a: number, b: number) => a !== b,
-    ],
-    [ts.SyntaxKind.LessThanToken, (a: number, b: number) => a < b],
-    [ts.SyntaxKind.LessThanEqualsToken, (a: number, b: number) => a <= b],
-    [ts.SyntaxKind.GreaterThanToken, (a: number, b: number) => a > b],
-    [ts.SyntaxKind.GreaterThanEqualsToken, (a: number, b: number) => a >= b],
-]);
-
-const postFixUnaryExpression = new Map<ts.SyntaxKind, (a: number) => number>([
-    [ts.SyntaxKind.PlusPlusToken, (a: number) => a + 1],
-    [ts.SyntaxKind.MinusMinusToken, (a: number) => a - 1],
-]);
-
-const preFixUnaryExpression = new Map<ts.SyntaxKind, (a: number) => number>([
-    [ts.SyntaxKind.PlusPlusToken, (a: number) => a + 1],
-    [ts.SyntaxKind.MinusMinusToken, (a: number) => a - 1],
-    [ts.SyntaxKind.PlusToken, (a: number) => +a],
-    [ts.SyntaxKind.MinusToken, (a: number) => -a],
-]);
-
-const booleanOperations = new Map<
-    ts.SyntaxKind,
-    (a: boolean, b: boolean) => boolean
->([
-    [
-        ts.SyntaxKind.EqualsEqualsEqualsToken,
-        (a: boolean, b: boolean) => a === b,
-    ],
-    [
-        ts.SyntaxKind.ExclamationEqualsEqualsToken,
-        (a: boolean, b: boolean) => a !== b,
-    ],
-]);
-
-function ApplyBinaryOperation(
-    opToken: ts.BinaryExpression["operatorToken"],
-    left: number | boolean,
-    right: number | boolean
-) {
-    const regularOperation = operations.get(opToken.kind);
-    const numberBooleanOperation = numberBooleanOperations.get(opToken.kind);
-    const booleanOperation = booleanOperations.get(opToken.kind);
-
-    if (typeof left === "number" && typeof right === "number") {
-        if (regularOperation !== undefined) {
-            return regularOperation(left, right);
-        } else if (numberBooleanOperation !== undefined) {
-            return numberBooleanOperation(left, right);
-        } else {
-            throw new Error("Can't process this Binary Expression");
-        }
-    } else if (typeof left === "boolean" && typeof right === "boolean") {
-        if (booleanOperation !== undefined) {
-            return booleanOperation(left, right);
-        } else {
-            throw new Error("Can't process this Binary Expression");
-        }
-    } else {
-        return undefined;
-    }
-}
-function editVariables(
-    node: ts.ExpressionStatement,
-    sourceFile: ts.SourceFile,
-    detectedVariableStatements: VariableStatementAnalysis[]
-): VariableStatementAnalysis[] | undefined {
-    const nodeExpression = node.expression;
-    //calculate binary expression and update its value
-    if (ts.isBinaryExpression(nodeExpression)) {
-        const elementIndex = detectedVariableStatements.findIndex(
-            (variables) => {
-                return variables.name === nodeExpression.left.getText();
-            }
-        );
-        //variable not found in the array
-        if (elementIndex === -1) {
-            throw new Error("Variable not defined");
-        }
-
-        if (detectedVariableStatements[elementIndex].variableType === "const") {
-            throw new Error("Can't redefine a constant");
-        }
-
-        //Get the new variable value
-        const newVariableValue = processExpression(
-            nodeExpression.right,
-            detectedVariableStatements
-        );
-
-        const expressionLocation = getNodePosition(sourceFile, nodeExpression);
-
-        const identifierLocation = getNodePosition(
-            sourceFile,
-            nodeExpression.left
-        );
-
-        //Check if undefined and throw an error if so
-        if (newVariableValue === undefined) {
-            throw Error("Value is undefined");
-        }
-
-        //since right would always be binary expression we want to process that, and update the variable value
-
-        detectedVariableStatements.push({
-            name: detectedVariableStatements[elementIndex].name,
-            value: newVariableValue,
-            //TODO: add you cant change constants and so
-            variableType: detectedVariableStatements[elementIndex].variableType,
-            text: node.getText(),
-            expressionLocation,
-            identifierLocation,
-        });
-
-        return detectedVariableStatements;
-    } else if (
-        //else if we encounter a i++ or --i case
-        ts.isPostfixUnaryExpression(nodeExpression) ||
-        ts.isPrefixUnaryExpression(nodeExpression)
-    ) {
-        if (ts.isIdentifier(nodeExpression.operand)) {
-            const elementIndex = detectedVariableStatements.findIndex(
-                (variables) => {
-                    return variables.name === nodeExpression.operand.getText();
-                }
-            );
-
-            //variable not found
-            if (elementIndex === -1) {
-                throw new Error("Variable not defined");
-            }
-
-            if (
-                detectedVariableStatements[elementIndex].variableType ===
-                "const"
-            ) {
-                throw new Error("Can't redefine a constant");
-            }
-
-            const operation = postFixUnaryExpression.get(
-                nodeExpression.operator
-            );
-
-            const value = detectedVariableStatements[elementIndex].value;
-
-            const expressionLocation = getNodePosition(
-                sourceFile,
-                nodeExpression
-            );
-
-            const identifierLocation = getNodePosition(
-                sourceFile,
-                nodeExpression.operand
-            );
-
-            //value has to be a number because you can't do True++
-            if (operation !== undefined && typeof value === "number") {
-                detectedVariableStatements.push({
-                    name: detectedVariableStatements[elementIndex].name,
-
-                    value: operation(value),
-
-                    text: node.getText(),
-
-                    variableType:
-                        detectedVariableStatements[elementIndex].variableType,
-
-                    expressionLocation,
-                    identifierLocation,
-                });
-                return detectedVariableStatements;
-            } else {
-                throw new Error("Operation is Undefined");
-            }
-        } else {
-            //Case where --5/5-- or ++7/7++ which can't be calculated
-            throw new Error(
-                "The operand of an increment or decrement operator must be a variable or a property"
-            );
-        }
-    }
-    return undefined;
-}
-
-function processExpression(
-    node:
-        | ts.Expression
-        | ts.NumericLiteral
-        | ts.Identifier
-        | ts.ParenthesizedExpression
-        | ts.PrefixUnaryExpression
-        | undefined,
-    detectedVariableStatements: VariableStatementAnalysis[]
-): number | boolean | undefined {
-    if (node === undefined) {
-        throw new Error("Expression is undefined");
-    }
-
-    if (ts.isBinaryExpression(node)) {
-        const left = processExpression(node.left, detectedVariableStatements);
-        const right = processExpression(node.right, detectedVariableStatements);
-
-        if (left !== undefined && right !== undefined) {
-            return ApplyBinaryOperation(node.operatorToken, left, right);
-        }
-        return undefined;
-    } else if (ts.isNumericLiteral(node)) {
-        //in case variables were defined just with one numeric literal for example: const x = 2;
-        return parseFloat(node.getText());
-    } else if (ts.isIdentifier(node)) {
-        const identifierValue = detectedVariableStatements.find((variables) => {
-            return variables.name === node.getText();
-        });
-        if (identifierValue === undefined) {
-            throw new Error(
-                "Identifier" +
-                    node.getText() +
-                    " cannot be found or undefined, please define a variable before using it"
-            );
-        }
-
-        return identifierValue.value;
-    } else if (ts.isParenthesizedExpression(node)) {
-        //in case we encounter a (...) situation, for example const a = 2 + (2 + 4)
-        return processExpression(node.expression, detectedVariableStatements);
-    } else if (ts.isPrefixUnaryExpression(node)) {
-        if (ts.isIdentifier(node.operand)) {
-            //Case where we are doing ++i or --i
-            const elementIndex = detectedVariableStatements.findIndex(
-                (variables) => {
-                    return variables.name === node.operand.getText();
-                }
-            );
-            //variable not found
-            if (elementIndex === -1) {
-                throw new Error(
-                    "Variable: " + node.operand.getText + " not defined"
-                );
-            }
-
-            //Get operation from the map and apply it
-            const operation = preFixUnaryExpression.get(node.operator);
-            const value = detectedVariableStatements[elementIndex].value;
-
-            //Apply the correct operation
-            if (operation !== undefined && typeof value === "number") {
-                return operation(value);
-            }
-            return undefined;
-        } else {
-            if (
-                node.operator === ts.SyntaxKind.PlusPlusToken ||
-                node.operator === ts.SyntaxKind.MinusMinusToken
-            ) {
-                //Can't do --4 or ++3 when defining a variable
-                throw new Error(
-                    "The operand of an increment or decrement operator must be a variable or a property"
-                );
-            }
-
-            //Cases where its -2 or +4 which is okay.
-            const value = processExpression(
-                node.operand,
-                detectedVariableStatements
-            );
-
-            //Get operation from the map and apply it
-            const operation = preFixUnaryExpression.get(node.operator);
-            if (typeof value === "number" && operation !== undefined) {
-                return operation(value);
-            } else {
-                return undefined;
-            }
-        }
-    } else if (node.kind === ts.SyntaxKind.TrueKeyword) {
-        return true;
-    } else if (node.kind === ts.SyntaxKind.FalseKeyword) {
-        return false;
-    } else {
-        throw new Error("Cannot process this expression: " + node.getText());
-    }
 }
 
 /**
